tests/wpt/mozilla/tests/mozilla/document-compatMode-strict.html

<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script>
            test(function() {
                assert_equals(document.compatMode, "CSS1Compat", "test1-0, Strict HTML");
            });
        </script>
    </head>
</html>
